Move lookup functions from `InstanceData` to `ComponentInstance`
More prep for the removal of `InstanceData`.
